requirements.txt
streamlit
pandas
numpy
matplotlib
pillow
altair
scikit-learn
tensorflow
vega-datasets
scipy
google-generativeai
google-cloud
google-auth